Tests/LibWeb/Text/input/css/animated-background-image-timer-doesnt-stop-after-full-layout-tree-teardown.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    #target {
        width: 100px;
        height: 50px;
        background-image: url("../wpt-import/images/anim-gr.gif");
    }
</style>
<div id="target"></div>
<dialog id="dialog">Hello</dialog>
<script>
    const imageURL = "../wpt-import/images/anim-gr.gif";

    asyncTest(async done => {
        await new Promise(resolve => window.addEventListener("load", resolve));

        internals.dumpLayoutTree(document);
        println(`timer active before teardown: ${internals.imageAnimationStateForURL(imageURL).timerActive}`);

        dialog.showModal();
        internals.dumpLayoutTree(document);
        println(`timer active after teardown: ${internals.imageAnimationStateForURL(imageURL).timerActive}`);

        target.style.display = "none";
        internals.dumpLayoutTree(document);
        println(`timer active after hiding: ${internals.imageAnimationStateForURL(imageURL).timerActive}`);
        done();
    });
</script>
