Tests/LibWeb/Text/expected/css/animated-background-image-timer-doesnt-stop-after-full-layout-tree-teardown.txt

timer active before teardown: true
timer active after teardown: true
timer active after hiding: true
